integ: move integ into its own "subcrate"
