--- tests/bench.rs
@@ -1,1 +1,0 @@
-
