+++ tests/src/lib.rs
@@ -1,0 +1,5 @@
+extern crate pty;
+extern crate tempdir;
+
+mod harness;
+mod integ;
